fix: update mobile and tablet styles for total funds cards
